@@ -9,22 +9,24 @@
   "main": "js/app.js",
   "private": true,
   "dependencies": {
+    "babel-cli": "^6.26.0",
     "babel-plugin-transform-runtime": "^6.4.3",
     "classnames": "^2.2.3",
-    "esp-js": "^1.1.3",
-    "esp-js-devtools": "0.0.8",
-    "react": "^15.3.0",
-    "react-dom": "^15.3.0"
+    "esp-js": "^2.0.0",
+    "esp-js-devtools": "0.0.15",
+    "prop-types": "^15.6.1",
+    "react": "16.4.0",
+    "react-dom": "16.4.0"
   },
   "devDependencies": {
     "babel-core": "^6.4.5",
-    "babel-loader": "^6.2.1",
+    "babel-loader": "7.1.4",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-preset-es2015": "^6.3.13",
-    "babel-preset-react": "^6.3.13",
+    "babel-preset-react": "^6.24.1",
     "babel-preset-stage-0": "^6.3.13",
-    "webpack": "^1.12.12",
-    "webpack-dev-server": "^1.14.1"
+    "webpack": "4.12.0",
+    "webpack-dev-server": "3.1.4"
   },
   "scripts": {
     "start": "webpack-dev-server --inline --hot --watch --progress --colors --config webpack-config.js"
